% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma-sep-to-matrix.R
\name{comma2mat}
\alias{comma2mat}
\title{Convert comma-separated lists of selections to a multiple-response questionnaire to a list of variables}
\usage{
comma2mat(x, name = "Q", sep = ",")
}
\arguments{
\item{x}{variable to convert}

\item{sep}{(default ',') separator between numbers}

\item{(default}{'Q') name for new variable, which will be followed by numbers to identify each question.}
}
\description{
Converts a variable containing list of selections, e.g. 1,4,10 to a list of variables, e.g. V1, V2, ..., V10, containing TRUE of FALSE indicating whether the corresponding item was selected. BUG: The function may fail mysteriously if one of the 'inputs' is not a number between 1 and the maximum number.
}
\examples{
df <- data.frame(id = 1:4,
     ans = c('1,2,3','5,2,1,4','',NA))
df
df <- cbind(df,comma2mat(df$ans,"Ques."))
df    
}

