% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haven.R
\name{merge_check}
\alias{merge_check}
\title{merge_check: Check compatibility of variables before merging.}
\usage{
merge_check(data1, data2)
}
\arguments{
\item{data1,}{data2 two data frames to compare.}
}
\value{
a data frame with variable names and classes in each input data frame.
}
\description{
Flags variables with the same name but different classes in two
data frames to be merged.
}

