% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haven.R
\name{toR}
\alias{toR}
\title{toR: Transform files read from SPSS.}
\usage{
toR(data)
}
\arguments{
\item{data}{a data frame created by reading a SPSS file with
\code{\link[haven]{read_sav}} in the 'haven' package.}
}
\value{
a data frame with all 'labelled' and 'character' variables
transformed to factors.
}
\description{
Transforms variables in files read with \code{\link[haven]{read_sav}} into
variables common in R.
}
\examples{
\dontrun{
library(haven)
dat <- read_sav("SPSSfile.sav")
dat <- toR(dat)
write_sav(dat, "OUT.sav")
}
}

