% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haven.R
\name{vallab}
\alias{vallab}
\title{vallab: Read and modify SPSS value labels}
\usage{
vallab(x, ...)
}
\arguments{
\item{x}{a variable which may be have a 'label' attribute}
}
\value{
the labels attribute of objects of class 'labelled'
}
\description{
Read and modify value labels for individual variables in a data frame read by \code{\link[haven]{read_sav}} in the 'haven' package.
}
\examples{
\dontrun{
dat <- read_sav("SPSSfile.sav")
vallab(dat$Gender)
vallab(dat$Gender) <- c('Male','Female')  # changing from, say, 'M' and 'F'
write_sav(dat,"OUT.sav")
}
}

