% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haven.R
\name{varlab}
\alias{varlab}
\title{varlab: Read and modify variable labels}
\usage{
varlab(x, ...)
}
\arguments{
\item{x}{a variable which may be have a 'label' attribute}
}
\value{
the label attribute
}
\description{
Read and modify variable labels for individual variables in a data frame read by \code{\link[haven]{read_sav}} in the 'haven' package.
}

