% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven.R
\name{varlabels}
\alias{varlabels}
\alias{varlabels<-}
\title{varlabels: extract and restore SPSS variable labels}
\usage{
varlabels(data)
}
\arguments{
\item{data}{a data frame, usually read with \code{\link[haven]{read_sav}}.}
}
\value{
a named list of variable labels
}
\description{
Extract SPSS variable labels from a data frame read with 
\code{\link[haven]{read_sav}} in the 'haven' package.
}
\details{
Many manipulations of data frames in R lose the \code{label} attribute
that holds SPSS variable labels. 'varlab' and 'varlab<-' allow the
variable labels to be extracted and later restored.
}
\examples{
\dontrun{
library(haven)
dat <- read_sav("SPSSfile.sav")
varlabs <- varlabels(dat)
dat <- toR(dat)
# work with dat in R
varlabels(dat) <- varlabs
write_sav(dat, "OUT.sav") # should have variable labels
} 
}

